refactor(renamed file, added type to props): renamed file and added types to props
