+++ src/app/page.tsx
@@ -1,14 +1,14 @@
 "use client";
 
 import React, { useState } from "react";
-import Flow1 from "@/components/container/flow1";
-import Flow2 from "@/components/container/flow2";
-import Flow3 from "@/components/container/flow3";
 import FooterLinks from "@/components/container/footer";
 import AuthXLogo from "@/components/container/authXLogo";
-import Flow4Auth from "@/components/container/flow4-auth";
-import Flow5Auth from "@/components/container/flow5-auth";
 import AuthOptions from "@/components/container/authOptions";
+import Flow1 from "@/components/container/create-account-flow/flow1";
+import Flow2 from "@/components/container/create-account-flow/flow2";
+import Flow3 from "@/components/container/create-account-flow/flow3";
+import Flow4Auth from "@/components/container/create-account-flow/flow4-auth";
+import Flow5Auth from "@/components/container/create-account-flow/flow5-auth";
 
 const Login = () => {
   const [createAccount, setCreateAccout] = useState(false);
@@ -20,11 +20,11 @@ const Login = () => {
   const [currentStep, setCurrentStep] = useState(1);
 
   const handleNextStep = () => {
-    setCurrentStep(currentStep + 1);
+    setCurrentStep((currentStep) => currentStep + 1);
   };
 
   const handlePreviousStep = () => {
-    setCurrentStep(currentStep - 1);
+    setCurrentStep((currentStep) => currentStep - 1);
   };
 
   return (
